third_party/WebKit/LayoutTests/fast/css/pseudo-in-range-expected.txt

Tests that we find controls if they have a range limitation and are in-range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

==> :in-range should match to an INPUT without a value:
PASS document.querySelector("input[type=number]:in-range").id is "number1"


==> :in-range should match to an INPUT with an in-range value:
PASS document.querySelector("input[type=number]:in-range").id is "number1"
PASS document.querySelector("input[type=range]:in-range").id is "range1"
PASS document.querySelectorAll(":in-range").length is 2
PASS successfullyParsed is true

TEST COMPLETE

 
